fixed bug in vfs and kvfs

+++ include/string.h
@@ -3,7 +3,7 @@
 
 #include <am.h>
 
-void *memset(void *b, int c, size_t n);
+void *memset(void *b, int c, int n);
 void *memcpy(void *dest, const void *src, size_t n);
 size_t strlen(const char* s);
 char *strcpy(char *dest, const char *src);
