+++ src/nslibc/string.c
@@ -1,6 +1,6 @@
 #include <am.h>
 
-void *memset(void *b, int c, size_t n) {
+void *memset(void *b, int c, int n) {
     char realc = c;
     char *B = (char*)b;
     while (--n >= 0) {
